+++ src/database/models/user.js
@@ -1,21 +1,16 @@
+'use strict';
 import { Model } from 'sequelize';
 
 export default (sequelize, DataTypes) => {
   class User extends Model {
-    /**
-     * Helper method for defining associations.
-     * This method is not a part of Sequelize lifecycle.
-     * The `models/index` file will call this method automatically.
-     */
     static associate(models) {
-      // define association here
-      User.HasOne(models.ROLES, {
+      User.hasOne(models.ROLES, {
         foreignKey: 'roleId',
         as: 'role',
         onDelete: 'CASCADE',
       });
 
-      User.HasOne(models.Carts, {
+      User.hasOne(models.Carts, {
         foreignKey: 'User_id',
         as: 'cart',
         onDelete: 'CASCADE',
